@@ -3,9 +3,9 @@ import 'package:google_fonts/google_fonts.dart';
 
 class modified_text extends StatelessWidget {
   final String text;
-  final Color color;
-  final double size;
-  const modified_text({Key key, this.text, this.color, this.size}) : super(key: key);
+  final Color? color;
+  final double? size;
+  const modified_text({Key? key,required this.text, this.color, this.size}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
