@@ -4,8 +4,8 @@ import 'package:movieapp/description.dart';
 import '../utils/text.dart';
 
 class TopRated extends StatelessWidget {
-  final List toprated;
-  const TopRated({Key key, this.toprated}) : super(key: key);
+  final List? toprated;
+  const TopRated({Key? key, this.toprated}) : super(key: key);
 
 
 
@@ -20,18 +20,18 @@ class TopRated extends StatelessWidget {
             modified_text(text: 'Top Rated', size: 26,),
             SizedBox(height: 10,),
             Container(height: 270,
-                child: ListView.builder(itemCount: toprated.length, scrollDirection: Axis.horizontal ,itemBuilder: (context, index){
+                child: ListView.builder(itemCount: toprated?.length, scrollDirection: Axis.horizontal ,itemBuilder: (context, index){
                   return InkWell(
                     onTap: (){
-                      Navigator.push(context, MaterialPageRoute(builder: (context)=>Description(name: toprated[index]['title'],
-                        bannerurl: 'https://image.tmdb.org/t/p/w500'+toprated[index]['backdrop_path'],
-                        posterurl: 'https://image.tmdb.org/t/p/w500'+toprated[index]['poster_path'],
-                        description: toprated[index]['overview'],
-                        vote: toprated[index]['vote_average'].toString(),
-                        launch_on: toprated[index]['release_date'],)));
+                      Navigator.push(context, MaterialPageRoute(builder: (context)=>Description(name: toprated?[index]['title'],
+                        bannerurl: 'https://image.tmdb.org/t/p/w500'+toprated?[index]['backdrop_path'],
+                        posterurl: 'https://image.tmdb.org/t/p/w500'+toprated?[index]['poster_path'],
+                        description: toprated?[index]['overview'],
+                        vote: toprated![index]['vote_average'].toString(),
+                        launch_on: toprated?[index]['release_date'],)));
 
                     },
-                    child: toprated[index]['title']!=null?Container(
+                    child: toprated?[index]['title']!=null?Container(
                       padding: EdgeInsets.all(5),
                       width: 250,
                       child: Column(
@@ -43,13 +43,13 @@ class TopRated extends StatelessWidget {
                                 borderRadius: BorderRadius.circular(10),
                                 image: DecorationImage(
                                 image: NetworkImage(
-                                    'https://image.tmdb.org/t/p/w500' + toprated[index]['backdrop_path']
+                                    'https://image.tmdb.org/t/p/w500' + toprated?[index]['backdrop_path']
                                 ), fit: BoxFit.cover
                             )),
                           ),
                           SizedBox(height: 10,),
-                          Container(child: modified_text(text: toprated[index]['title'] != null?
-                          toprated[index]['title']:'Loading' ),)
+                          Container(child: modified_text(text: toprated?[index]['title'] != null?
+                          toprated![index]['title']:'Loading' ),)
                         ],
                       ),
                     ):Container(),
